# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/audit": {
      "get": {
        "operationId": "AuditController_getLogs",
        "parameters": [
          {
            "name": "entity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "ISO date or YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "ISO date or YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get audit logs with optional filters (admin only)",
        "tags": [
          "Audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "SwimSync API",
    "description": "REST API for SwimSync services",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      }
    },
...
  }
}